imxrt10xx: fix LED blinking
The LED write routine is meant to take a value between 0 for off and 255 for full brightness.
the current 10xx port treats 0 as off, and anything else as on.  This creates a case where the
LED is on virtually all the time, except for a tiny blink off every 256 timer intervals.

(this is for non-pwm, simple LEDs)

This fix makes (value >= 128) is ON and (value < 128) off, and the LED looks right.