boards/olimex_rt1010/flash_config.c: Fix flash settings.
The settings were not the right ones. Instead of the W25Q16 flash
listed initially in the old documentation, the production boards use
now an EN25Q16B-104HIP chip, which needs different arguments to enable
Quad mode. With the existing arguments, the flash may get locked.

It can then be unlocked using the flash-sfdp mode for uploading the UF2
bootloader.

Signed-off-by: robert-hh <robert@hammelrath.com>